2 Player Yacht Dice
Was planning on making an RL agent for this, turns out its waaaaay harder than I expected. This is a whole paper on how hard it is: https://web.stanford.edu/class/aa228/reports/2018/final75.pdf. But I had made this already when I found out, so now its on github.